Port Boosts & Eggs to JSON
