+++ EGG9000.Common/Helpers/EggIncStatics.cs
@@ -1,0 +1,72 @@
+﻿using EGG9000.Bot;
+using EGG9000.Common.JsonData.EiStatics;
+using Humanizer;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text.RegularExpressions;
+
+namespace EGG9000.Common.Helpers {
+    public class EggIncStatics {
+        public static EggIncEgg GetEggById(Ei.Egg egg) {
+            return GetEggById((int)egg);
+        }
+        public static EggIncEgg GetEggById(int id) {
+            return Root.Get().eggIncEggs.FirstOrDefault(x => x.id == id);
+        }
+
+        private static string ToStartCase(string input) {
+            var reg1 = new Regex(@"_");
+            var reg2 = new Regex(@"(?: |\b)(\w)");
+            input = reg1.Replace(input, " ");
+            input = reg2.Replace(input, match => match.Value.ToUpper());
+            return input;
+        }
+
+        public static EggIncBoost FromId(string id) {
+            return Root.Get().eggIncBoosts.FirstOrDefault(x => x.id == id);
+        }
+
+        public static string GetReward(Ei.Contract.Types.Goal goal) {
+            switch (goal.RewardType) {
+                case Ei.RewardType.EggsOfProphecy:
+                    return $"<:Egg_of_Prophecy:669981330477547580> {goal.RewardAmount}";
+                case Ei.RewardType.PiggyFill:
+                    return $"<:Piggy_bank:724396277676113955> + {goal.RewardAmount.ToEggString()}";
+                case Ei.RewardType.Gold:
+                    return $"<:Egg_Golden:692439755798872075>  {goal.RewardAmount.ToEggString()}";
+                case Ei.RewardType.EpicResearchItem:
+                    var researchItem = (EpicResearchCalc.GetEpicResearchConfig().AsQueryable().FirstOrDefault(x => x.Id == goal.RewardSubType.ToLower()));
+                    var goldenEggRefund = (int)(researchItem?.Costs?.Last() * goal?.RewardAmount);
+                    var goldenEggRefundString = $" (<:Golden_Egg_GE:692439755798872075> {(goldenEggRefund >= 1000 ? (goldenEggRefund / 1000).ToString() + 'K' : goldenEggRefund)})";
+                    return $"{researchItem?.Title ?? ToStartCase(goal.RewardSubType)} +{goal.RewardAmount}{(goldenEggRefund == default ? "" : goldenEggRefundString)}";
+                case Ei.RewardType.Boost:
+                    var boost = FromId(goal.RewardSubType);
+
+                    var type = (object)boost.type switch {
+                        EggIncBoostTypeEnum.Earnings => $"{boost.emoji} {boost.value}x for {boost.timeMinutes.Humanize().ShortenTime()}",
+                        EggIncBoostTypeEnum.BoostBeacon => $"{boost.emoji} {boost.value}x for {boost.timeMinutes.Humanize().ShortenTime()}",
+                        EggIncBoostTypeEnum.InternalHatchery => $"{boost.emoji} {boost.value}x for {boost.timeMinutes.Humanize().ShortenTime()}",
+                        EggIncBoostTypeEnum.FarmValue => $"{boost.emoji} {boost.value / 100}x of Farm Value",
+                        EggIncBoostTypeEnum.SoulEggs => $"{boost.emoji} {boost.value}x for {boost.timeMinutes.Humanize().ShortenTime()}",
+                        EggIncBoostTypeEnum.SoulMirror => $"{boost.emoji} for {boost.timeMinutes.Humanize().ShortenTime()}",
+                        EggIncBoostTypeEnum.UnlimitedHatchery => $"{boost.emoji} for {boost.timeMinutes.Humanize().ShortenTime()}",
+                        _ => $"Unknown Boost Type: {goal.RewardSubType}",
+                    };
+                    return (goal.RewardAmount > 1 ? goal.RewardAmount.ToString() : "") + type;
+                case Ei.RewardType.SoulEggs:
+                    return $"<:Egg_soul:724341890794913964> {goal.RewardAmount.ToEggString()}";
+                case Ei.RewardType.PiggyMultiplier:
+                    return $"<:Piggy_bank:724396277676113955> {goal.RewardAmount}x";
+                case Ei.RewardType.PiggyLevelBump:
+                    return $"<:Piggy_up:812846540426838016> Level + {goal.RewardAmount}";
+                case Ei.RewardType.ArtifactCase:
+                    return $"<:Afx_reward:877681508607987772> {goal.RewardAmount}";
+                case Ei.RewardType.ShellScript:
+                    return $"<:tickets:998630687831769189> {goal.RewardAmount}";
+                default:
+                    return goal.RewardType.ToString();
+            }
+
+        }
+    }
+}
